+++ include/srtp_priv.h
@@ -141,8 +141,8 @@ typedef struct srtp_stream_ctx_t_ {
     srtp_sec_serv_t rtcp_services;
     direction_t direction;
     bool allow_repeat_tx;
-    int *enc_xtn_hdr;
-    int enc_xtn_hdr_count;
+    uint8_t *enc_xtn_hdr;
+    size_t enc_xtn_hdr_count;
     uint32_t pending_roc;
 } strp_stream_ctx_t_;
